+++ Source/NSPortCoder.m
@@ -1215,7 +1215,7 @@ static IMP	_xRefImp;	/* Serialize a crossref.	*/
 	  return;
 	}
 
-      GSIMapAddPair(_cIdMap, (GSIMapKey)anObject, (GSIMapVal)0);
+      GSIMapAddPair(_cIdMap, (GSIMapKey)anObject, (GSIMapVal)(NSUInteger)0);
     }
   else if (anObject == nil)
     {
@@ -1292,7 +1292,8 @@ static IMP	_xRefImp;	/* Serialize a crossref.	*/
 	       *	and add it to the map of unconditionay encoded ones.
 	       */
 	      GSIMapRemoveKey(_cIdMap, (GSIMapKey)anObject);
-	      GSIMapAddPair(_uIdMap, (GSIMapKey)anObject, (GSIMapVal)0);
+	      GSIMapAddPair(_uIdMap,
+		(GSIMapKey)anObject, (GSIMapVal)(NSUInteger)0);
 	      [anObject encodeWithCoder: self];
 	    }
 	  return;
@@ -1306,7 +1307,7 @@ static IMP	_xRefImp;	/* Serialize a crossref.	*/
 	  if (node == 0)
 	    {
 	      node = GSIMapAddPair(_uIdMap,
-		(GSIMapKey)anObject, (GSIMapVal)++_xRefO);
+		(GSIMapKey)anObject, (GSIMapVal)(NSUInteger)++_xRefO);
 	    }
 	  else
 	    {
@@ -1459,7 +1460,7 @@ static IMP	_xRefImp;	/* Serialize a crossref.	*/
 		if (node == 0)
 		  {
 		    GSIMapAddPair(_ptrMap,
-			(GSIMapKey)*(void**)buf, (GSIMapVal)0);
+		      (GSIMapKey)*(void**)buf, (GSIMapVal)(NSUInteger)0);
 		    type++;
 		    buf = *(char**)buf;
 		    (*_eValImp)(self, eValSel, type, buf);
@@ -1473,7 +1474,7 @@ static IMP	_xRefImp;	/* Serialize a crossref.	*/
 		if (node == 0)
 		  {
 		    node = GSIMapAddPair(_ptrMap,
-			(GSIMapKey)*(void**)buf, (GSIMapVal)++_xRefP);
+		      (GSIMapKey)*(void**)buf, (GSIMapVal)(NSUInteger)++_xRefP);
 		  }
 		else
 		  {
@@ -1539,7 +1540,7 @@ static IMP	_xRefImp;	/* Serialize a crossref.	*/
 				format: @"negative class version"];
 		  }
 		node = GSIMapAddPair(_clsMap,
-			(GSIMapKey)(void*)c, (GSIMapVal)++_xRefC);
+		  (GSIMapKey)(void*)c, (GSIMapVal)(NSUInteger)++_xRefC);
 		/*
 		 *	Encode tag and crossref number.
 		 */
@@ -1589,7 +1590,7 @@ static IMP	_xRefImp;	/* Serialize a crossref.	*/
 	    if (node == 0)
 	      {
 		node = GSIMapAddPair(_ptrMap,
-			(GSIMapKey)(void*)s, (GSIMapVal)++_xRefP);
+		  (GSIMapKey)(void*)s, (GSIMapVal)(NSUInteger)++_xRefP);
 		(*_xRefImp)(_dst, xRefSel, _GSC_SEL, node->value.nsu);
 		/*
 		 *	Encode selector.
@@ -1620,7 +1621,7 @@ static IMP	_xRefImp;	/* Serialize a crossref.	*/
 	    if (node == 0)
 	      {
 		node = GSIMapAddPair(_ptrMap,
-			(GSIMapKey)*(char**)buf, (GSIMapVal)++_xRefP);
+		  (GSIMapKey)*(char**)buf, (GSIMapVal)(NSUInteger)++_xRefP);
 		(*_xRefImp)(_dst, xRefSel, _GSC_CHARPTR, node->value.nsu);
 		(*_eSerImp)(_dst, eSerSel, buf, type, nil);
 	      }
